+++ tests/cli_tests/zcnsc/client/authorizer.go
