+++ tests/cli_tests/zcnsc/client/authorizers.go
