+++ tests/cli_tests/zcnsc/client/balance.go
