+++ tests/cli_tests/zcnsc/client/chainstatetest.go
