+++ tests/cli_tests/zcnsc/client/mint.go
