+++ tests/cli_tests/zcnsc/client/models.go
