+++ tests/cli_tests/zcnsc/client/setup.go
