+++ tests/cli_tests/zcnsc/client/smart_contract_const.go
