+++ tests/cli_tests/zcnsc/client/status.go
